requirement.txt
nfcpy==1.0.3
